Add an "accessibility time factor" to provide for extra time on timed tests.
This accesibility time factor is a user property and is set when editing
a student on the "Accounts Manager" page. The time that a student will
have to complete any timed test is the product of the "Test Time Limit"
set for the test on the "Set Detail" page, and this accesibility time
factor for the student taking the test.  By default the accesibility
time factor for each student is 1, but can be set to something like 1.5
to determine that a student is allowed time and a half to complete timed
tests.

The point of this is that it is a bit tedious to need to go through all
timed tests and change the test time limit for all of the students that
need to be given extra time for accesibility accomodations. With this
you only need to set one number, and it is rather convenient to do so
for all of the students in the class that need it from one page.

Since this is a per user setting, this requires the addition of a new
column to the user table in the database.