Revert number inputs for fields with labels on the problem set detail page.
This is a quick fix that will make things function for now, and is good
for a hotfix.  This just reverts all of the fields that were switched to
number inputs and that have labels for special negative values back to
text inputs and the functionality prior to WeBWorK 2.20.

The fields that do not have those labels can still be number inputs.
That will work fine for those.

Also set the default for showMeAnother to the "Course Default" instead
of "Never" which is what it should be.

This resolves issue #2820 for the current release.

We can try to do better for the next release, but I don't see any other
easy way to fix this that is appropriate for a hotfix.