When a course is renamed update the LTI course map for the course.
This deletes the course map entry for the course's previous name, and
then creates a new entry for the course's new name.

Since the `renameCourse` method is also used when unarchiving a course
to a different name, care needs to be taken to ensure that this is only
done for a course rename, and not for that situation. Thus there is a
new `updateLTICourseMap` option for the method.  This is only true when
the method is called to rename a course directly.

This addresses issue #2824.