Rework error handling when a problem fails to render in a test.
First, fix a general issue when a problem fails to render that results
in a message that has no relation to the actual problem. The issue is
that the `create_ans_str_from_responses` method of the
`WeBWorK::Utils::ProblemProcessing` module is called with a `$pg` hash
that doesn't contain the right things, and the method is not safe
guarded against that. This fixes issue #2829.

Second, rework how error handling is done in tests.  Instead of just
rendering the errors for problems that failed to render, still render
all problems on the page.  Show the errors for the problems that failed,
but still render all of the problems that didn't fail.  This means, that
the problem navigation is still shown.   In fact the entire page renders
normally, but the error message is shown for the problems that fail to
render.  So the student can at least continue to work the rest of the
test normally.  The student can even grade the test on a page for which
a problem fails to render.

Note that the "request information" is no longer shown for these
messages when in a test.  A special `briefErrorOutput` stash value can
be set when the `templates/ContentGenerator/Base/error_output.html.ep`
template is rendered to skip that. Although, I really thing that
information should never be shown in the page for any exception that
occurs. I don't think that the time, method, uri, and HTTP headers of
the request, ever help with these exceptions.