Make MathJax show errors for bad TeX but only in the problem editor.
There have been requests to either remove this extension or at least
make it so that those editing problems do not have it loaded, as it
makes it easier to determine what is wrong with TeX in a problem.

This pull request makes it so that these errors are only shown in the PG
problem editor. Actually they are shown any time the renderRPC endpoint
is used and the `showMathJaxErrors` parameter is set to a true value,
but the only time that webwork2 now does this is in the problem editor.

This still includes thhe change from the `webwork_url` to the
`webwork_js_config` method in the `WeBWorK::ContentGenerator` module.
There is also a `webwork_url` method in the `Mojolicious::WeBWorK`
module that is already available for all controller modules (since it is
a Mojolicious helper method), and having this other one overrides that
one and it is confusing to have both that return almost the same value.
The only difference is that `WeBWorK::ContentGenerator` method called
the `location` helper which returns the empty string if the root URL is
'/', and the `webwork_url` helper returns '/' in that case. I don't know
what I was thinking creating the `WeBWorK::ContentGenerator` method
which was really just an alies for the `location` helper method anyway.