Add a standalone variant to check_latex.
The packages and TikZ libraries that are used by pg are all tested
(including the TikZ spath library that is going to be used once
https://github.com/openwebwork/pg/pull/1336 is merged).

Note that the "active" option was removed from the preview package in
the check_latex_article.tex and check_latex_exam.tex files because that
suppresses output.  It is enough to test that the preview package loads,
and this makes the more important things that are being tested more like
they are for actual hardcopy generation and not for equation generation.
We really need to remove images mode.