Add an option to add headers to only SSL request responses.
The option is `extra_ssl_headers` in the `conf/webwork2.mojolicious.yml`
file.  It works just like the `extra_headers` option, except that the
headers are only added to responses to secure requests.

This is to address a need to add the `Strict-Transport-Security` header
to SSL request responses that was brought up in the forums.  See
https://forums.openwebwork.org/mod/forum/discuss.php?d=8782#p22468. That
header should not be added to non SSL requests.