Fix the openwebwork.org link in hardcopy.
The `webwork2.sty` file currently creates this link using the `\url`
command.  That command attempts to detect the link type.  In this case
that results in a file link.  Depending on the PDF viewer used, that
results in different behavior.  In Firefox's PDF viewer it behaves as if
it is not a link at all, but just text.  Clicking on it does nothing. In
Google Chrome's PDF viewer or Evince it is a link, but clicking on it
results in a message being displayed that the file does not exist.

So this switches to using the `\href` command instead. That requires
that both the link URL and link text explicitly be given, but reliably
results in a working link.