Tweak sticky answers with problem randomization a bit more.
Previously if a new problem version is being opened, then the answers
from a previous version were deleted from the form fields of the last
form submission.

Now, the form fields form any previous form submission are simply not
even sent to PG.  If this is a new problem version, there is absolutely
nothing in the form fields that PG needs. This is a much more efficient
and thorough approach.  It ensures that if a problem uses some other
inputs that are not part of the answer, those inputs are cleared as
well.  GeoGebra problems do this for instance to save their state.