Make the "Include OPL" and "Include Contrib" check states in the library browser persistent.
The status of the checks is saved to local storage and updated when the
page loads.  So whatever state they had the last time the page was open
is restored. Note that the settings are saved per user id and course.
So you can set the checks differently for each course.

This was (essentially) asked for in issue #2857.