Configure rpc endpoints to disable cookies for specific courses/users.
Add option to allow the `allow_unsecured_rpc` option to be a hash
of course ids, which can be set to 0 or 1 to disable cookies only
for that course. Further, the course id could also be a hash of
user ids that can be set to 0 or 1 to disable cookies only for the
specified user id.