Allow viewing achievement items when acting as another user.
Instead of greying out the "Use Achievement Reward" button when
acting as another user, which only lists the available achievement
items to use, grey out the actual use button to let an instructor
be able to view which items a student could use on that assignment.

It is still not possible to actually use an achievement as another
user if that button was active, and further the form tags are also
removed from the list of items.

This address #2856.