Fix the MathQuill usage on the problem grader page.
When MathQuill was switched to a published npm package instead of being
used from the Github repository, the `node_modules` location changed.
This was all updated on the PG side of things, but I forgot to do it for
the problem grader (which is the only place MathQuill is used for
webwork2).