Improvements for the problem renderer on the problem grader page.
First, this makes it so that feedback is shown on initial render.  This
occurs both on initial page load (for the initially selected user), and
also occurs when a different user is selected from the dropdown.

Second, this fixes some issues with tests and answers.  The test answer
prefix needs to be preserved in the form so that checking answers works.
Also, the `ProblemGrader.html.ep` template was constructing the quiz
prefix from the problem number and should be using the `problem_id`
(which is what the GatewayQuiz.pm module uses).  That means the last
answers were not working at all for tests with random problem order.