Configurable feedback email links show problem grader.
This adds a new course configuration option that allows the instructor
to choose if the problem grader will be open or not when the links sent
in a feedback email from a problem are followed.

I am finding that to be rather obtrusive, and not something that I ever
want.  I know I am not alone on this.