Make the problem graders also save the sub_status when needed.
Basically, anytime that reduced scoring is not enabled or if it is but
it is before the reduced scoring period, then the sub_status needs to
also be set to the same thing as the status.  Otherwise the "unreduced"
score will not be computed correctly.  So the problem graders now take
this into account.

This fixes issue #2873.